git-svn-id: svn://svn.icms.temple.edu/lammps-ro/trunk@9802 f3b2605a-c512-4ea7-a41b-209d697bcdaa

+++ src/KSPACE/pppm.cpp
@@ -1357,7 +1357,7 @@ void PPPM::compute_gf_denom()
     gf_b[0] = 4.0 * (gf_b[0]*(l-m)*(l-m-0.5));
   }
 
-  int ifact = 1;
+  bigint ifact = 1;
   for (k = 1; k < 2*order; k++) ifact *= k;
   double gaminv = 1.0/ifact;
   for (l = 0; l < order; l++) gf_b[l] *= gaminv;
